+++ src/common/classes/typed-message.class.js
@@ -1,5 +1,5 @@
 /*
- * /src/common/classes/typed-message.class.js
+ * pwix:typed-message/src/common/classes/typed-message.class.js
  *
  * A class which implements the ITypedMessage interface.
  */
@@ -8,11 +8,13 @@ import _ from 'lodash';
 const assert = require( 'assert' ).strict; // up to nodejs v16.x
 import mix from '@vestergaard-company/js-mixin';
 
-import { OrdStack } from 'meteor/pwix:orderable-stack';
+import { OStack } from 'meteor/pwix:orderable-stack';
 
 import { ITypedMessage } from '../interfaces/ityped-message.iface.js';
 
-export class TypedMessage  extends  mix( OrdStack.Orderable ).with( ITypedMessage ){
+import { TypeOrder } from '../definitions/type-order.def.js';
+
+export class TypedMessage  extends  mix( OStack.Orderable ).with( ITypedMessage ){
 
     // static data
 
@@ -33,9 +35,12 @@ export class TypedMessage  extends  mix( OrdStack.Orderable ).with( ITypedMessag
         return this;
     }
 
-    // the method each IOrderable implementation must provde
+    // the method each IOrderable implementation must provide
     IOrderableCompare( a, b ){
-        let res = 0;
-        return res;
+        assert( a instanceof ITypedMessage, 'expected an ITypedMessage instance' );
+        assert( b instanceof ITypedMessage, 'expected an ITypedMessage instance' );
+        const level_a = a.ITypedMessageType();
+        const level_b = b.ITypedMessageType();
+        return TypeOrder.compare( level_a, level_b );
     }
 }