+++ src/common/js/index.js
@@ -15,9 +15,9 @@ import { TypedMessage } from '../classes/typed-message.class';
 
 import './global.js';
 import './constants.js';
+import './i18n.js';
 //
 import './configure.js';
-import './i18n.js';
 
 TM.MessageType = MessageType;
 TM.TypeOrder = TypeOrder;