+++ src/server/js/index.js
@@ -4,5 +4,4 @@
 
 import '../../common/js/index.js';
 
-//import './check_npms.js';
-//import './collections.js';
+import './check_npms.js';